# api-auth-server/src/main/resources/static/api.yml
openapi: 3.0.3
info:
  title: OAuth2 Authorization Service API
  description: Provides API interfaces for OAuth2 authorization services, including standard OAuth2 interfaces and migration tool interfaces
  version: 1.0.0

servers:
  - url: 'http://localhost:9000'
    description: Local Development Server

tags:
  - name: oauth2
    description: Standard OAuth2 Interface
  - name: legacy
    description: Compatible with the old version of OAuth2 Interface
  - name: migration
    description: Data Migration Tool Interface
  - name: admin
    description: Administrator Interface

paths:
  /oauth2/token:
    post:
      tags:
        - oauth2
      summary: Get OAuth2 Access Token
      description: Standard OAuth2 Token Endpoint for obtaining access tokens
      operationId: getToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum: [client_credentials]
                  description: Authorization Type
                client_id:
                  type: string
                  description: Client ID
                client_secret:
                  type: string
                  description: Client Secret
                scope:
                  type: string
                  description: Requested Scope of Permissions
              required:
                - grant_type
                - client_id
                - client_secret
      responses:
        '200':
          description: Token Obtained Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
                  scope:
                    type: string
        '400':
          description: Bad Request
        '401':
          description: Authentication Failed

  /oauth2/introspect:
    post:
      tags:
        - oauth2
      summary: Token Introspection Endpoint
      description: Used to verify the validity of the token and obtain token information
      operationId: introspectToken
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Token to be checked
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Token Type Hint
              required:
                - token
      responses:
        '200':
          description: Token Introspection Result
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                  client_id:
                    type: string
                  username:
                    type: string
                  scope:
                    type: string
                  exp:
                    type: integer
                  iat:
                    type: integer

  /oauth2/jwks:
    get:
      tags:
        - oauth2
      summary: JWK Key Set Endpoint
      description: Provides public keys for verifying JWT tokens
      operationId: getJwks
      responses:
        '200':
          description: Successfully Obtained JWK Set
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        kid:
                          type: string
                        kty:
                          type: string
                        alg:
                          type: string
                        use:
                          type: string
                        n:
                          type: string
                        e:
                          type: string

  /oauth/token:
    post:
      tags:
        - legacy
      summary: Old Version Token Endpoint
      description: Token acquisition endpoint compatible with the old version of OAuth2
      operationId: getLegacyToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token Obtained Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
        '401':
          description: Authentication Failed

  /oauth/token_key:
    get:
      tags:
        - legacy
      summary: Old Version Public Key Endpoint
      description: Public key acquisition endpoint compatible with the old version of OAuth2
      operationId: getLegacyTokenKey
      responses:
        '200':
          description: Successfully Obtained Public Key
          content:
            application/json:
              schema:
                type: object
                properties:
                  alg:
                    type: string
                  value:
                    type: string
                    description: PEM format public key

  /api/admin/migration/clients:
    post:
      tags:
        - migration
        - admin
      summary: Migrate Client Data
      description: Trigger the migration operation of OAuth2 client data
      operationId: migrateClients
      security:
        - basicAuth: []
      responses:
        '200':
          description: Migration Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  message:
                    type: string
        '500':
          description: Migration Failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [error]
                  message:
                    type: string

  /api/admin/migration/tokens:
    post:
      tags:
        - migration
        - admin
      summary: Migrate Token Data
      description: Trigger the migration operation of OAuth2 access tokens
      operationId: migrateTokens
      security:
        - basicAuth: []
      responses:
        '200':
          description: Migration Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationResponse'
        '500':
          description: Migration Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/migration/all:
    post:
      tags:
        - migration
        - admin
      summary: Migrate All Data
      description: Trigger the migration operation of all OAuth2 related data
      operationId: migrateAll
      security:
        - basicAuth: []
      responses:
        '200':
          description: Migration Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationResponse'
        '500':
          description: Migration Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /clients:
    post:
      tags:
        - admin
      summary: Create a New OAuth2 Client
      description: Create a new OAuth2 client
      operationId: createClient
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                  description: Client ID
                client_secret:
                  type: string
                  description: Client Secret
                resource_ids:
                  type: array
                  items:
                    type: string
                  description: Resource ID List
                scope:
                  type: array
                  items:
                    type: string
                  description: Scope of Permissions List
                grant_types:
                  type: array
                  items:
                    type: string
                  description: Authorization Type List
                expires_in:
                  type: integer
                  description: Token Expiration Time (seconds)
                additional_information:
                  type: object
                  description: Additional Client Information
              required:
                - client_id
                - client_secret
      responses:
        '201':
          description: Client Created Successfully
        '400':
          description: Bad Request
        '401':
          description: Authentication Failed
        '409':
          description: Client ID Already Exists

    get:
      tags:
        - admin
      summary: Query Client List
      description: Query the client list based on the resourceId
      operationId: getClients
      security:
        - basicAuth: []
      parameters:
        - name: resource_id
          in: query
          schema:
            type: string
          description: Resource ID
      responses:
        '200':
          description: Successfully Obtained Client List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '401':
          description: Authentication Failed

  /clients/{client_id}:
    get:
      tags:
        - admin
      summary: Get Specific Client Information
      description: Query detailed information of the client based on the client_id
      operationId: getClient
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '200':
          description: Successfully Obtained Client Information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

    put:
      tags:
        - admin
      summary: Update Client Information
      description: Update the configuration information of the specified client
      operationId: updateClient
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scope:
                  type: array
                  items:
                    type: string
                grant_types:
                  type: array
                  items:
                    type: string
                expires_in:
                  type: integer
                additional_information:
                  type: object
                resource_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Client Updated Successfully
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

    delete:
      tags:
        - admin
      summary: Delete Client
      description: Delete the specified client
      operationId: deleteClient
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '204':
          description: Client Deleted Successfully
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

  /clients/{client_id}/secret:
    put:
      tags:
        - admin
      summary: Update Client Secret
      description: Update the secret of the specified client
      operationId: updateClientSecret
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_secret:
                  type: string
                  description: New Client Secret
              required:
                - client_secret
      responses:
        '200':
          description: Client Secret Updated Successfully
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

  /clients/{client_id}/tokens:
    get:
      tags:
        - admin
      summary: Get Client Token List
      description: Get the access token list of the specified client
      operationId: getClientTokens
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
        - name: scope
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by Scope of Permissions
      responses:
        '200':
          description: Successfully Obtained Token List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Token'
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

    delete:
      tags:
        - admin
      summary: Delete Client Token
      description: Delete the access token of the specified client
      operationId: deleteClientTokens
      security:
        - basicAuth: []
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Client ID
        - name: scope
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by Scope of Permissions
      responses:
        '204':
          description: Token Deleted Successfully
        '401':
          description: Authentication Failed
        '404':
          description: Client Not Found

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    TokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
        client_id:
          type: string
        client_secret:
          type: string
        scope:
          type: string
      required:
        - grant_type
        - client_id
        - client_secret

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
        scope:
          type: string

    MigrationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
      required:
        - status
        - message

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
      required:
        - status
        - message

    Client:
      type: object
      properties:
        client_id:
          type: string
        client_secret:
          type: string
        resource_ids:
          type: array
          items:
            type: string
        scope:
          type: array
          items:
            type: string
        grant_types:
          type: array
          items:
            type: string
        expires_in:
          type: integer
        additional_information:
          type: object
      required:
        - client_id
        - client_secret

    Token:
      type: object
      properties:
        token_id:
          type: string
        client_id:
          type: string
        scope:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - token_id
        - client_id